src/test/java/hgrcompany/hgrshop/MemberRepositoryTest.java
package hgrcompany.hgrshop;

import hgrcompany.hgrshop.domain.Member;
import org.assertj.core.api.Assertions;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class) // junit에 스프링 테스트임을 알려줌
@SpringBootTest
public class MemberRepositoryTest {

    @Autowired MemberRepository memberRepository;
    @Test
    @Transactional // test가 끝난 후 롤백
    @Rollback(false) // 테스트한 후 롤백을 하지 않는 어노테이션
    public void testMember() throws Exception {
        // given
        Member member = new Member();
        member.setName("hgr");
        // when
        Long savedId = memberRepository.save(member);
        Member findMember = memberRepository.find(savedId);

        //then
        Assertions.assertThat(findMember.getId()).isEqualTo(member.getId());
        Assertions.assertThat(findMember.getName()).isEqualTo(member.getName());


    }

}